rename sdk/agent to sdk/bench
